Added CSS for project layout
Created CSS file and added initial styling for the project layout. Imported the normalize.css library and the "Kumbh Sans" font from Google Fonts. Defined global styles for the page, including font, padding, margin, border, outline, text decoration, color, box sizing, and cursor. Styled the body element with flexbox properties to center and align content. Created a header with flexbox properties for a navigation section and a user section. Styled the navigation links with hover effects and a underline animation. Styled the user section with icons for a shopping cart and a user avatar. Created a main section with flexbox properties and added styling for product photos and small filtered photos.